Fix multiline strings
The multiline strings in `defaults/main.yml` were incorrectly formatted,
which was causing them to contain literal `"` characters and to have
erroneous `\n` characters appended to the end of them.

This has been fixed by removing the quote characters and switching to
the `>-` block specifier which prevents the addition of a final newline.